Added Site Template
So here is a basic template I've come up with that pretty closely models
the diagram that we submitted in the SDS. I have yet to add the
login/registration popup box that Megan implemented in here example (I'm
going to try an add that tomorrow). I also have not tested the search
bar (which will work on enter). Also there are no links as of yet (which
will require some modification of the CSS). This is the basic idea of
the site at least and hopefully it'll help get the ball rolling.